+++ src/Livewire/Settings/CustomFieldGroups/CustomFieldGroupShow.php
@@ -2,12 +2,14 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\Settings\CustomFieldGroups;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\FieldGroup;
 
 class CustomFieldGroupShow extends Component
 {
+    use AuthorizesRequests;
     use Toast;
 
     public FieldGroup $fieldGroup;
@@ -15,6 +17,8 @@ class CustomFieldGroupShow extends Component
     public function delete($id)
     {
         if ($fieldGroup = FieldGroup::find($id)) {
+            $this->authorize('delete', $fieldGroup);
+
             $fieldGroup->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.custom_field_group_deleted')), redirectTo: route('laravel-crm.field-groups.index'));
