+++ src/Livewire/Settings/Labels/LabelShow.php
@@ -2,12 +2,14 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\Settings\Labels;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\Label;
 
 class LabelShow extends Component
 {
+    use AuthorizesRequests;
     use Toast;
 
     public Label $label;
@@ -15,6 +17,8 @@ class LabelShow extends Component
     public function delete($id)
     {
         if ($label = Label::find($id)) {
+            $this->authorize('delete', $label);
+
             $label->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.label_deleted')), redirectTo: route('laravel-crm.labels.index'));
