+++ src/Livewire/Settings/PipelineStages/PipelineStageShow.php
@@ -2,12 +2,14 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\Settings\PipelineStages;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\PipelineStage;
 
 class PipelineStageShow extends Component
 {
+    use AuthorizesRequests;
     use Toast;
 
     public PipelineStage $pipelineStage;
@@ -15,6 +17,8 @@ class PipelineStageShow extends Component
     public function delete($id)
     {
         if ($pipelineStage = PipelineStage::find($id)) {
+            $this->authorize('delete', $pipelineStage);
+
             $pipelineStage->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.pipeline_stage_deleted')), redirectTo: route('laravel-crm.pipeline-stages.index'));
