+++ src/Livewire/Settings/ProductCategories/ProductCategoryShow.php
@@ -2,12 +2,14 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\Settings\ProductCategories;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\ProductCategory;
 
 class ProductCategoryShow extends Component
 {
+    use AuthorizesRequests;
     use Toast;
 
     public ProductCategory $productCategory;
@@ -15,6 +17,8 @@ class ProductCategoryShow extends Component
     public function delete($id)
     {
         if ($productCategory = ProductCategory::find($id)) {
+            $this->authorize('delete', $productCategory);
+
             $productCategory->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.product_category_deleted')), redirectTo: route('laravel-crm.product-categories.index'));
